refactor auction specifc migration code to legacy v0_17 pkg
