+++ x/auction/legacy/v0_16/codec.go
@@ -1,0 +1,16 @@
+package types
+
+import (
+	types "github.com/cosmos/cosmos-sdk/codec/types"
+	v017auction "github.com/kava-labs/kava/x/auction/types"
+)
+
+func RegisterInterfaces(registry types.InterfaceRegistry) {
+	registry.RegisterInterface(
+		"kava.auction.v1beta1.GenesisAuction",
+		(*v017auction.GenesisAuction)(nil),
+		&v017auction.SurplusAuction{},
+		&v017auction.DebtAuction{},
+		&v017auction.CollateralAuction{},
+	)
+}
